clean up folders and project structure
