+++ apps/studio/src/routes/editor/LayersPanel/Tree/TreeRow.tsx
@@ -1,16 +1,20 @@
-import { forwardRef } from 'react';
 import type { LayerNode } from '@onlook/models/element';
-import type { RowRendererProps } from 'react-arborist';
+import type { RefObject } from 'react';
+import type { NodeApi } from 'react-arborist';
 
-const TreeRow = forwardRef<HTMLDivElement, RowRendererProps<LayerNode>>((props, ref) => {
-    const { attrs, children } = props;
+interface TreeRowProps {
+    node: NodeApi<LayerNode>;
+    innerRef: RefObject<HTMLDivElement>;
+    attrs: { [key: string]: any };
+    children: React.ReactNode;
+}
+
+const TreeRow = ({ innerRef, attrs, children }: TreeRowProps) => {
     return (
-        <div ref={ref} {...attrs} className="outline-none">
+        <div ref={innerRef} {...attrs} className="outline-none">
             {children}
         </div>
     );
-});
-
-TreeRow.displayName = 'TreeRow';
+};
 
 export default TreeRow;
